Refactoring package structure and Setup pytest with coverage
