+++ tests/test_basic.py
@@ -3,12 +3,12 @@ Author: nitin
 Date: 18/7/17
 Description: 
 """
-from decorators import cache
-from backends import RedisBackend
 import time
 from math import factorial
 
-from utils import args_to_key, hash_key
+from cache4py.decorators import cache
+from cache4py.storage.redis import RedisBackend
+from cache4py.utils import hash_key
 
 redis_backend = RedisBackend('localhost', 6379)
 
@@ -17,23 +17,27 @@ redis_backend = RedisBackend('localhost', 6379)
 def redis_target_function(x):
     return factorial(x)
 
+
 def uncached_target_function(x):
     return factorial(x)
 
+
 def test_redis():
-    print("Start test")
+    print("Start redis caching test")
+
     start_time = time.time()
     for i in range(5):
-      result = uncached_target_function(75000)
-    first_elapsed_time = time.time() - start_time
+      _ = uncached_target_function(75000)
+    uncached_time = time.time() - start_time
+
     start_time = time.time()
     for i in range(5):
-      result = redis_target_function(75000)
-    second_elapsed_time = time.time() - start_time
-    print("Time difference: before: {0}, after: {1}".format(first_elapsed_time, second_elapsed_time))
+      _ = redis_target_function(75000)
+    cached_time = time.time() - start_time
+
+    print("Time difference: before: {0}, after: {1}".format(uncached_time, cached_time))
+
+    assert(cached_time < uncached_time)
+
     hashed_key = hash_key(75000)
     redis_backend.delete(key_name=hashed_key)
-
-if __name__ == '__main__':
-    print('aaya')
-    test_redis()
